Extract `calculateLineParamsFromPoints` from `useLines`
